requirements.txt
streamlit==1.32.0
python-dotenv>=1.0.0 
Pillow==10.3.0
PyPDF2==3.0.1
langchain==0.3.1
langchain-community==0.3.1
langchain-google-genai==0.7.2
google-generativeai==0.7.2
faiss-cpu
googletrans==4.0.0-rc1
duckduckgo-search==8.1.1
SpeechRecognition==3.14.3
pyaudio==0.2.14
deep-translator==1.11.4
gtts==2.5.4
langchain_experimental==0.3.2
langchain_core==0.3.6
google-cloud-texttospeech==2.27.0
google-cloud-speech==2.33.0
streamlit-mic-recorder==0.0.8